requirements.txt
alembic==1.12.0
attrs==23.2.0
blinker==1.8.2
beautifulsoup4==4.12.3
click==8.1.7
flake8==5.0.4
Flask==3.0.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.3
importlib_metadata==7.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
lxml==5.3.0
Jinja2==3.1.4
Mako==1.3.3
MarkupSafe==2.1.5
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
py==1.11.0
pycodestyle==2.9.1
pyflakes==2.5.0
pytest==7.1.3
pytest-env==0.6.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==13.7 
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.21
tomli==2.0.1
typing_extensions==4.11.0
Werkzeug==3.0.0
WTForms==3.0.1
zipp==3.18.1
